+++ lib/doc-governor/index.ts
@@ -1,0 +1,3 @@
+export * from "./DocScanner";
+export * from "./DocGovernor";
+export * from "./DocRenderer";